Renamed module to Vertex
