+++ java/apirag-vertex/src/main/java/com/datasqrl/ai/models/vertex/VertexChatModel.java
@@ -1,4 +1,4 @@
-package com.datasqrl.ai.models.google;
+package com.datasqrl.ai.models.vertex;
 
 import com.datasqrl.ai.models.GenericLanguageModel;
 import lombok.AllArgsConstructor;
@@ -6,7 +6,7 @@ import lombok.Getter;
 
 @AllArgsConstructor
 @Getter
-public enum GoogleChatModel implements GenericLanguageModel {
+public enum VertexChatModel implements GenericLanguageModel {
 
   GEMINI_1_5_FLASH("gemini-1.5-flash", 1000000, 8192),
   GEMINI_1_5_PRO("gemini-1.5-pro", 1000000, 8192);
