+++ java/apirag-vertex/src/main/java/com/datasqrl/ai/models/vertex/VertexChatProviderFactory.java
@@ -1,4 +1,4 @@
-package com.datasqrl.ai.models.google;
+package com.datasqrl.ai.models.vertex;
 
 import com.datasqrl.ai.backend.FunctionBackend;
 import com.datasqrl.ai.models.AbstractChatProviderFactory;
@@ -9,9 +9,9 @@ import com.google.auto.service.AutoService;
 import org.apache.commons.configuration2.Configuration;
 
 @AutoService(ChatProviderFactory.class)
-public class GoogleChatProviderFactory extends AbstractChatProviderFactory {
+public class VertexChatProviderFactory extends AbstractChatProviderFactory {
 
-  public static final String PROVIDER_NAME = "google";
+  public static final String PROVIDER_NAME = "vertex";
   private static final String VERTEX_PROJECT_ID_KEY = "vertex-project-id";
   private static final String VERTEX_LOCATION_KEY = "vertex-location";
 
@@ -26,6 +26,6 @@ public class GoogleChatProviderFactory extends AbstractChatProviderFactory {
     ErrorHandling.checkArgument(vertexProjectId != null && !vertexProjectId.isBlank(), "Need to configure vertex-project-id.");
     String vertexLocation = modelConfiguration.getString(VERTEX_LOCATION_KEY);
     ErrorHandling.checkArgument(vertexLocation != null && !vertexLocation.isBlank(), "Need to configure vertex-location.");
-    return new GoogleChatProvider(getModel(modelConfiguration, GoogleChatModel.class), vertexProjectId, vertexLocation, backend, prompt);
+    return new VertexChatProvider(getModel(modelConfiguration, VertexChatModel.class), vertexProjectId, vertexLocation, backend, prompt);
   }
 }
