+++ java/pom.xml
@@ -14,9 +14,9 @@
     <module>apirag-openai</module>
     <module>apirag-groq</module>
     <module>apirag-bedrock</module>
+    <module>apirag-vertex</module>
     <module>apirag-graphql</module>
     <module>apirag-rest</module>
-    <module>apirag-google</module>
     <module>apirag-server</module>
   </modules>
